uint32

+++ core/graphBolt/KickStarterEngine.h
@@ -107,10 +107,10 @@ public:
   template <class T> struct DependencyData {
     uintV parent;
     T value;
-    uint16_t level;
+    uint32_t level;
     DependencyData() : level(MAX_LEVEL), value(), parent(MAX_PARENT) {}
 
-    DependencyData(uint16_t _level, T _value, uint32_t _parent)
+    DependencyData(uint32_t _level, T _value, uint32_t _parent)
         : level(_level), value(_value), parent(_parent) {}
 
     DependencyData(const DependencyData &object)
